+++ app/src/main/java/com/tokyoc/line_client/SettingActivity.kt
@@ -1,0 +1,58 @@
+package com.tokyoc.line_client
+
+import android.content.DialogInterface
+import android.content.Intent
+import android.os.Bundle
+import android.support.v7.app.AlertDialog
+import android.support.v7.app.AppCompatActivity
+import android.util.Log
+import android.view.KeyEvent
+import android.widget.Button
+import android.widget.ListView
+import android.widget.Toast
+import com.google.firebase.auth.FirebaseAuth
+import io.realm.Realm
+
+class SettingActivity : AppCompatActivity() {
+    private lateinit var realm: Realm
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_setting)
+
+        val token = intent.getStringExtra("token")
+
+        //戻るボタンを押した時の処理
+        findViewById<Button>(R.id.return_button).setOnClickListener {
+            val intent = Intent(this, GroupActivity::class.java)
+            intent.putExtra("token", token)
+            startActivity(intent)
+        }
+
+        //サインアウトボタンを押した時の処理
+        findViewById<Button>(R.id.signout_button).setOnClickListener {
+            val intent = Intent(this, SigninActivity::class.java)
+            AlertDialog.Builder(this).apply {
+                setTitle("Sign Out")
+                setMessage("Really Sign Out?")
+                setPositiveButton("OK", DialogInterface.OnClickListener { _, _ ->
+                    FirebaseAuth.getInstance().signOut()
+                    startActivity(intent)
+                })
+                setNegativeButton("Cancel", null)
+                show()
+            }
+        }
+    }
+
+    override fun onKeyDown(keyCode: Int, event: KeyEvent): Boolean {
+        when (keyCode) {
+            KeyEvent.KEYCODE_BACK -> {
+                Toast.makeText(applicationContext, "this key is invalid", Toast.LENGTH_LONG)
+                        .show()
+                return false
+            }
+            else -> return super.onKeyDown(keyCode, event)
+        }
+    }
+}
