+++ app/src/main/res/layout/activity_group.xml
@@ -6,35 +6,49 @@
     android:orientation="vertical"
     android:gravity="center">
 
-    <ListView
-        android:id="@+id/group_list_view"
-        android:layout_width="match_parent"
-        android:layout_height="550dp"/>
-    
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
         android:gravity="center_horizontal">
 
-        <Button
-            android:id="@+id/signout_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Sign out"/>
+        <!--<Button-->
+            <!--android:id="@+id/signout_button"-->
+            <!--android:layout_width="wrap_content"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:text="Sign out"/>-->
 
         <Button
             android:id="@+id/member_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="友達一覧"/>
+            android:text="友達"/>
 
         <Button
-            android:id="@+id/make_group_button"
+            android:id="@+id/group_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="New Group"/>
+            android:text="グループ"/>
+
+        <Button
+            android:id="@+id/setting_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="設定" />
 
     </LinearLayout>
+
+    <ListView
+        android:id="@+id/group_list_view"
+        android:layout_width="match_parent"
+        android:layout_height="500dp"/>
+
+    <Button
+        android:id="@+id/make_group_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="グループ追加" />
+    
+
 
 </LinearLayout>
