+++ app/src/main/res/layout/activity_setting.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <Button
+        android:id="@+id/return_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="戻る" />
+
+    <Button
+    android:id="@+id/signout_button"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="Sign out"/>
+
+</LinearLayout>
